fix: adjust date range picker future allowed date for every company
