+++ apps/gauzy/src/app/@shared/report/report-base/report-base.component.ts
@@ -2,14 +2,14 @@ import { Component } from '@angular/core';
 import { combineLatest, debounceTime } from 'rxjs';
 import { Subject } from 'rxjs';
 import { filter, tap } from 'rxjs/operators';
-import * as moment from 'moment';
-import { IOrganization, ISelectedDateRange, ITimeLogFilters } from '@gauzy/contracts';
+import { IOrganization, ITimeLogFilters } from '@gauzy/contracts';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { TranslateService } from '@ngx-translate/core';
 import { distinctUntilChange, toUTC } from '@gauzy/common-angular';
 import { pick } from 'underscore';
 import { Store } from '../../../@core/services/store.service';
 import { TranslationBaseComponent } from '../../language-base/translation-base.component';
+import { getAdjustDateRangeFutureAllowed } from '../../../@theme/components/header/selectors/date-range-picker';
 
 @UntilDestroy({ checkProperties: true })
 @Component({
@@ -62,7 +62,7 @@ export class ReportBaseComponent extends TranslationBaseComponent {
 	}
 
 	getFilterRequest(request: ITimeLogFilters): ITimeLogFilters {
-		const { startDate, endDate } = this.getAdjustDateRangeFutureAllowed(request);
+		const { startDate, endDate } = getAdjustDateRangeFutureAllowed(request);
 		const { id: organizationId } = this.organization;
 		const { tenantId } = this.store.user;
 
@@ -76,42 +76,5 @@ export class ReportBaseComponent extends TranslationBaseComponent {
 			endDate: toUTC(endDate).format('YYYY-MM-DD HH:mm')
 		};
 		return filterRequest;
-	}
-
-	/**
-	 * We are having issue, when organization not allowed future date
-	 * When someone run timer for today, all statistic not displaying correctly
-	 *
-	 * @returns
-	 */
-	protected getAdjustDateRangeFutureAllowed(request: ITimeLogFilters): ISelectedDateRange {
-		const now = moment();
-		let { startDate, endDate } = request;
-		/**
-		 * If, user selected single day date range.
-		 */
-		if (
-			moment(moment(startDate).format('YYYY-MM-DD')).isSame(
-				moment(endDate).format('YYYY-MM-DD')
-			)
-		) {
-			startDate = moment(startDate).startOf('day').utc().toDate();
-			endDate = moment(endDate).endOf('day').utc().toDate();
-		}
-
-		/**
-		 * If, user selected TODAY date range.
-		 */
-		if (
-			moment(now.format('YYYY-MM-DD')).isSame(
-				moment(endDate).format('YYYY-MM-DD')
-			)
-		) {
-			endDate = moment.utc().toDate();
-		}
-		return {
-			startDate: moment(startDate).toDate(),
-			endDate: moment(endDate).toDate()
-		} as ISelectedDateRange
 	}
 }
