+++ apps/gauzy/src/app/@theme/components/header/selectors/date-range-picker/index.ts
@@ -1,2 +1,3 @@
 export * from './date-range-picker.component';
 export * from './date-range-picker.module';
+export * from './date-range-picker.setting';
