+++ apps/gauzy/src/app/pages/reports/amounts-owed-report/amounts-owed-report/amounts-owed-report.component.ts
@@ -15,6 +15,7 @@ import { TimesheetService } from './../../../../@shared/timesheet/timesheet.serv
 import { ReportBaseComponent } from './../../../../@shared/report/report-base/report-base.component';
 import { IChartData } from './../../../../@shared/report/charts/line-chart/line-chart.component';
 import { ChartUtil } from './../../../../@shared/report/charts/line-chart/chart-utils';
+import { getAdjustDateRangeFutureAllowed } from './../../../../@theme/components/header/selectors/date-range-picker';
 
 @UntilDestroy({ checkProperties: true })
 @Component({
@@ -59,7 +60,7 @@ export class AmountsOwedReportComponent extends ReportBaseComponent
 		this.filters = Object.assign(
 			{},
 			this.logRequest,
-			this.getAdjustDateRangeFutureAllowed(this.logRequest)
+			getAdjustDateRangeFutureAllowed(this.logRequest)
 		);
 		this.subject$.next(true);
 	}
