+++ apps/gauzy/src/app/pages/reports/payment-report/payment-report/payment-report.component.ts
@@ -14,6 +14,7 @@ import { ReportBaseComponent } from './../../../../@shared/report/report-base/re
 import { IChartData } from './../../../../@shared/report/charts/line-chart/line-chart.component';
 import { ChartUtil } from './../../../../@shared/report/charts/line-chart/chart-utils';
 import { PaymentService, Store } from './../../../../@core/services';
+import { getAdjustDateRangeFutureAllowed } from './../../../../@theme/components/header/selectors/date-range-picker';
 
 @UntilDestroy({ checkProperties: true })
 @Component({
@@ -58,7 +59,7 @@ export class PaymentReportComponent extends ReportBaseComponent
 		this.filters = Object.assign(
 			{},
 			this.logRequest,
-			this.getAdjustDateRangeFutureAllowed(this.logRequest)
+			getAdjustDateRangeFutureAllowed(this.logRequest)
 		);
 		this.subject$.next(true);
 	}
