+++ apps/gauzy/src/app/pages/reports/time-reports/time-reports/time-reports.component.ts
@@ -22,6 +22,7 @@ import { TimesheetService } from './../../../../@shared/timesheet/timesheet.serv
 import { ReportBaseComponent } from './../../../../@shared/report/report-base/report-base.component';
 import { ChartUtil } from './../../../../@shared/report/charts/line-chart/chart-utils';
 import { IChartData } from './../../../../@shared/report/charts/line-chart/line-chart.component';
+import { getAdjustDateRangeFutureAllowed } from './../../../../@theme/components/header/selectors/date-range-picker';
 
 @UntilDestroy({ checkProperties: true })
 @Component({
@@ -66,7 +67,7 @@ export class TimeReportsComponent extends ReportBaseComponent
 		this.filters = Object.assign(
 			{},
 			this.logRequest,
-			this.getAdjustDateRangeFutureAllowed(this.logRequest)
+			getAdjustDateRangeFutureAllowed(this.logRequest)
 		);
 		this.subject$.next(true);
 	}
